+++ packages/components/grid/src/filter-column.ts
@@ -3,7 +3,7 @@ import { getNameByPath, getValueByPath } from '@sl-design-system/shared';
 import { localized, msg } from '@lit/localize';
 import { html } from 'lit';
 import { property, state } from 'lit/decorators.js';
-import { GridColumn } from './column.js';
+import { GridColumn, GridColumnEvent } from './column.js';
 import { GridFilter } from './filter.js';
 
 export type GridFilterMode = 'select' | 'text';
@@ -13,8 +13,15 @@ export interface GridFilterOption {
   value?: unknown;
 }
 
+export class GridFilterValueChangeEvent<T> extends GridColumnEvent<T> {
+  constructor(column: GridColumn<T>, public readonly value: string | string[] | undefined) {
+    super('sl-filter-value-change', column);
+  }
+}
+
 @localized()
-export class GridFilterColumn extends GridColumn {
+// eslint-disable-next-line @typescript-eslint/no-explicit-any
+export class GridFilterColumn<T = any> extends GridColumn<T> {
   /** The internal options if none are provided. */
   @state() internalOptions?: GridFilterOption[];